refactor(report): move dataframe surface onto EvaluationReport, remove aggregation package
